Adding JRE download and backup command
Therefore we don't need JRE preinstalled now, but the user (should - untested) get an error message, which says that JRE is missing.
So there are two options:
* Installing JRE via AppCenter like before
* Installing JRE via service script using downloadJava

Additionally I added an backup command, which easily creates a tar.gz of the distribution and a timestamp in its archive name into distribution/backups.

Signed-off-by: Thomas Karl Pietrowski thopiekar@gmail.com (github: thopiekar)